+++ index.html
@@ -17,7 +17,7 @@
       src="https://kit.fontawesome.com/729d9295e6.js"
       crossorigin="anonymous"
     ></script>
-    <link rel="stylesheet" type="text/css" href="main.css" />
+    <link rel="stylesheet" type="text/css" href="./assets/styles/main.css" />
   </head>
 
   <body class="text-center">
@@ -62,15 +62,16 @@
     <div id="book-container" class="col-12">
       <div class="row">
         <div id="store" class="col-6"></div>
+        <div id="cart-modal"></div>
         <div id="cart" class="col-6"></div>
       </div>
     </div>
 
-
     <script
-  src="https://code.jquery.com/jquery-3.5.1.min.js"
-  integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0="
-  crossorigin="anonymous"></script>
+      src="https://code.jquery.com/jquery-3.5.1.min.js"
+      integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0="
+      crossorigin="anonymous"
+    ></script>
     <script
       src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js"
       integrity="sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN"
@@ -81,6 +82,6 @@
       integrity="sha384-B4gt1jrGC7Jh4AgTPSdUtOBvfO8shuf57BaghqFfPlYxofvL8/KUEfYiJOMMV+rV"
       crossorigin="anonymous"
     ></script>
-    <script type="text/javascript" src="main.js"></script>
+    <script type="module" src="./javascripts/main.js"></script>
   </body>
 </html>
